+++ src/frontend/apps/e2e/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['conversations/playwright'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-    project: ['./tsconfig.json'],
-  },
-  ignorePatterns: ['node_modules'],
-};