+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to
 - ♻️(chat) refactor AIAgentService for readability and maintainability
 - 🚸(oidc) ignore case when fallback on email #281
 - ⬆️(back) update pillow, django-pydantic-field, pypdf
+- ♻️(front) migrate from ESLint 8 to ESLint 9 flat config
 
 ### Fixed